+++ Sourcery/Generating/Generator.swift
@@ -8,6 +8,7 @@ import Stencil
 
 enum Generator {
     static func generate(_ types: Types, template: Template, arguments: [String: NSObject] = [:]) throws -> String {
+        Log.info("Rendering template \(template.sourcePath)")
         return try template.render(types: types, arguments: arguments)
     }
 }