+++ Sourcery/Generating/Template/JavaScript/JavaScriptTemplate.swift
@@ -1,20 +1,6 @@
 import JavaScriptCore
 import PathKit
 
-struct JavaScriptTemplateError: Error, CustomStringConvertible {
-    var exception: JSValue?
-
-    init(_ exception: JSValue?) {
-        self.exception = exception
-    }
-
-    var description: String {
-        let message = exception?.forProperty("message").toString() ?? "no message"
-
-        return message
-    }
-}
-
 final class JavaScriptTemplate: Template {
 
     let sourcePath: Path
@@ -27,6 +13,12 @@ final class JavaScriptTemplate: Template {
         jsContext = context
     }
 
+    init(templateString: String) {
+        self.template = templateString
+        sourcePath = ""
+        jsContext = JSContext()
+    }
+
     func render(types: Types, arguments: [String : NSObject]) throws -> String {
         let context = TemplateContext(types: types, arguments: arguments)
 
@@ -34,7 +26,26 @@ final class JavaScriptTemplate: Template {
         let path = Bundle(for: JavaScriptTemplate.self).path(forResource: "ejsbundle", ofType: "js")!
         let ejs = try String(contentsOfFile: path, encoding: .utf8)
 
-        var error: JavaScriptTemplateError?
+        var error: Swift.Error?
+        jsContext.exceptionHandler = { _, exception in
+            error = exception?.toString() ?? "Unknown JavaScript error"
+        }
+        jsContext.setObject(template, forKeyedSubscript: "template" as NSString)
+        jsContext.setObject(sourcePath.lastComponent, forKeyedSubscript: "templateName" as NSString)
+        jsContext.setObject(context.jsContext, forKeyedSubscript: "templateContext" as NSString)
+
+        let valueForKey: @convention(block) (TypesCollection, String) -> Any? = { target, key in
+            do {
+                return try target.types(forKey: key)
+            } catch let _error {
+                error = _error
+                return nil
+            }
+        }
+        jsContext.setObject(valueForKey, forKeyedSubscript: "valueForKey" as NSString)
+        jsContext.evaluateScript("templateContext.types.implementing = new Proxy(templateContext.types.implementing, { get: valueForKey })")
+        jsContext.evaluateScript("templateContext.types.inheriting = new Proxy(templateContext.types.inheriting, { get: valueForKey })")
+        jsContext.evaluateScript("templateContext.types.based = new Proxy(templateContext.types.based, { get: valueForKey })")
 
         let include: @convention(block) (String) -> [String:String] = { [unowned self] requestedPath in
             let requestedPath = Path(requestedPath)
@@ -54,30 +65,16 @@ final class JavaScriptTemplate: Template {
 
             return templateDictionary
         }
-
-        let exceptionHandler: (JSContext?, JSValue?) -> Void = { context, exception in
-            error = JavaScriptTemplateError(exception)
-        }
-
-        jsContext.setObject(template, forKeyedSubscript: "template" as NSString)
-        jsContext.setObject(context.jsContext, forKeyedSubscript: "templateContext" as NSString)
         jsContext.setObject(include, forKeyedSubscript: "include" as NSString)
-        jsContext.setObject(sourcePath.lastComponent, forKeyedSubscript: "templateName" as NSString)
-        jsContext.exceptionHandler = exceptionHandler
+
         jsContext.evaluateScript("var window = this; \(ejs)")
 
         if let error = error {
-            throw error
+            throw "\(sourcePath): \(error)"
         }
 
-        let content: String
-        if let contentObject = jsContext.objectForKeyedSubscript("content"), contentObject.isString {
-            content = contentObject.toString()
-        } else {
-            content = ""
-        }
-
-        return content
+        let content = jsContext.objectForKeyedSubscript("content").toString()
+        return content ?? ""
     }
 
 }