+++ Sourcery/Sourcery.swift
@@ -140,13 +140,21 @@ class Sourcery {
 
     fileprivate func templates(from: [Path]) throws -> [Template] {
         return try templatePaths(from: from).map {
-            if $0.extension == "swifttemplate" {
-                return try SwiftTemplate(path: $0)
-            } else if $0.extension == "ejs" {
-                return try JavaScriptTemplate(path: $0)
-            } else {
-                return try StencilTemplate(path: $0)
-            }
+            #if SWIFT_PACKAGE
+                if $0.extension == "swifttemplate" || $0.extension == "ejs" {
+                    throw "Swift and JavaScript templates are not supported when using Sourcery built with Swift Package Manager yet. Please use only Stencil templates. See https://github.com/krzysztofzablocki/Sourcery/issues/244 for details."
+                } else {
+                    return try StencilTemplate(path: $0)
+                }
+            #else
+                if $0.extension == "swifttemplate" {
+                    return try SwiftTemplate(path: $0)
+                } else if $0.extension == "ejs" {
+                    return try JavaScriptTemplate(path: $0)
+                } else {
+                    return try StencilTemplate(path: $0)
+                }
+            #endif
         }
     }
 
@@ -208,7 +216,7 @@ extension Sourcery {
                     return result == .approved
                 }
                 .map {
-                    try FileParser(verbose: verbose, contents: $0.contents, path: $0.path, module: modules?[index])
+                    try FileParser(contents: $0.contents, path: $0.path, module: modules?[index])
             }
 
             var previousUpdate = 0
@@ -237,7 +245,7 @@ extension Sourcery {
         }
 
         //! All files have been scanned, time to join extensions with base class
-        let types = Composer(verbose: verbose).uniqueTypes(parserResult)
+        let types = Composer().uniqueTypes(parserResult)
 
         track("Found \(types.count) types.")
         return (Types(types: types), inlineRanges)
@@ -278,7 +286,7 @@ extension Sourcery {
                                   unarchivedResult = unarchived
                               }
                           }, catch: { _ in
-            self.track("Failed to unarchive \(artifacts) due to error, re-parsing")
+            Log.warning("Failed to unarchive \(artifacts) due to error, re-parsing")
         }, finallyBlock: {})
 
         return unarchivedResult
@@ -313,10 +321,10 @@ extension Sourcery {
                 try writeIfChanged(Sourcery.generationHeader + result, to: outputPath)
             } else {
                 if prune && outputPath.exists {
-                    track("Removing \(outputPath) as it is empty.")
+                    Log.info("Removing \(outputPath) as it is empty.")
                     do { try outputPath.delete() } catch { track("\(error)") }
                 } else {
-                    track("Skipping \(outputPath) as it is empty.")
+                    Log.info("Skipping \(outputPath) as it is empty.")
                 }
             }
         }