+++ SourceryTests/Generating/JavaScriptTemplateSpecs.swift
@@ -40,6 +40,24 @@ class JavaScriptTemplateTests: QuickSpec {
                 let result = (try? (outputDir + Sourcery().generatedPath(for: templatePath)).read(.utf8))
                 expect(result).to(equal(expectedResult))
             }
+
+            it("rethrows template parsing errors") {
+                expect {
+                    try Generator.generate(Types(types: []), template: JavaScriptTemplate(templateString: "<% invalid %>"))
+                    }
+                    .to(throwError(closure: { (error) in
+                        expect("\(error)").to(equal(": ReferenceError: ejs:1\n >> 1| <% invalid %>\n\nCan\'t find variable: invalid"))
+                    }))
+            }
+
+            it("rethrows template runtime errors") {
+                expect {
+                    try Generator.generate(Types(types: []), template: JavaScriptTemplate(templateString: "<%= types.implementing.Some %>"))
+                    }
+                    .to(throwError(closure: { (error) in
+                        expect("\(error)").to(equal(": Unknown type Some, should be used with `based`"))
+                    }))
+            }
         }
     }
 }