+++ CHANGELOG.md
@@ -6,10 +6,22 @@
 
 ### New Features
 
+- Paths in config file are now relative to config file path by default, absolute paths should start with `/`
+- Improved logging and error reporting, added `--quiet` CLI option, added runtime errors for using invalid types in `implementing` and `inheriting`
 - Added support for includes in EJS templates (for example: `<%- include('myTemplate.js') %>`)
-- Errors during EJS template compilation / parsing are reported
+- Add the `lowerFirst` filter for Stencil templates.
+
+### Internal changes
+
+- Improved `AutoCases.stencil` to use `let` instead of computed `var`
 
 ## 0.6.0
+
+### New Features
+
+- Added some convenience accessors for classic, static and instance methods, and types and contained types grouped by names
+
+## 0.6
 
 ### New Features